+++ package/src/presets/next/index.ts
@@ -1,3 +1,6 @@
 export { PLUGIN_NEXT } from "#/presets/next/plugin";
 export { nextPreset } from "#/presets/next/preset";
-export { RULES_NEXTJS } from "#/presets/next/rules/nextjs";
+export {
+    RULES_NEXTJS,
+    RULES_NEXTJS_CORRECTNESS,
+} from "#/presets/next/rules/nextjs";