+++ package/CHANGELOG.md
@@ -11,6 +11,7 @@
 - add support for `react/jsx-no-literals`
 - export `RULES_OXC_SUSPICIOUS` from `@apst/oxlint/presets/common`
 - export `RULES_NEXTJS` from `@apst/oxlint/presets/next`
+- export `RULES_NEXTJS_CORRECTNESS` from `@apst/oxlint/presets/next`
 
 ### What's Changed